Arm: Move implementation of operations under arm module
