+++ emu/src/cpu/arm/mod.rs
@@ -1,3 +1,4 @@
 pub mod alu_instruction;
 pub mod instructions;
 pub mod mode;
+pub mod operations;
