+++ emu/src/cpu/arm7tdmi.rs
@@ -20,7 +20,7 @@ use crate::cpu::thumb::mode::ThumbModeOpcode;
 use crate::memory::internal_memory::InternalMemory;
 use crate::memory::io_device::IoDevice;
 
-use super::flags::{Indexing, LoadStoreKind, Offsetting, OperandKind, ReadWriteKind};
+use super::flags::{LoadStoreKind, OperandKind, ReadWriteKind};
 use super::registers::Registers;
 
 pub const SIZE_OF_ARM_INSTRUCTION: u32 = 4;
@@ -566,132 +566,6 @@ impl Arm7tdmi {
         Some(SIZE_OF_THUMB_INSTRUCTION)
     }
 
-    fn branch_and_exchange(&mut self, register: usize) -> Option<u32> {
-        let mut rn = self.registers.register_at(register);
-        let state: CpuState = rn.get_bit(0).into();
-        self.cpsr.set_cpu_state(state);
-
-        match self.cpsr.cpu_state() {
-            CpuState::Thumb => rn.set_bit_off(0),
-            CpuState::Arm => {
-                rn.set_bit_off(0);
-                rn.set_bit_off(1);
-            }
-        }
-
-        self.registers.set_program_counter(rn);
-
-        None
-    }
-
-    fn half_word_data_transfer(&mut self, op_code: ArmModeOpcode) -> Option<u32> {
-        let indexing: Indexing = op_code.get_bit(24).into();
-        let offsetting: Offsetting = op_code.get_bit(23).into();
-        let write_back = op_code.get_bit(21);
-        let load_store: LoadStoreKind = op_code.get_bit(20).into();
-        let rn_base_register = op_code.get_bits(16..=19);
-        let rd_source_destination_register = op_code.get_bits(12..=15);
-        let transfer_type = HalfwordTransferType::from(op_code.get_bits(5..=6) as u8);
-
-        let operand_kind: OperandKind = op_code.get_bit(22).into();
-
-        let offset = match operand_kind {
-            OperandKind::Immediate => {
-                let immediate_offset_high = op_code.get_bits(8..=11);
-                let immediate_offset_low = op_code.get_bits(0..=3);
-                (immediate_offset_high << 4) | immediate_offset_low
-            }
-            OperandKind::Register => {
-                let rm: usize = op_code.get_bits(0..=3).try_into().unwrap();
-                self.registers.register_at(rm)
-            }
-        };
-
-        let mut address = self
-            .registers
-            .register_at(rn_base_register.try_into().unwrap());
-
-        if rn_base_register == REG_PROGRAM_COUNTER {
-            // prefetching
-            address = address.wrapping_add(8);
-
-            if write_back {
-                panic!("WriteBack should not be specified when using R15 as base register.");
-            }
-
-            if indexing == Indexing::Post {
-                panic!("Post indexing uses write back but we're using R15 as base register.
-                Documentation says that when using R15 as base register WB should not be used. What should we do?");
-            }
-        }
-
-        let effective = match offsetting {
-            Offsetting::Down => address.wrapping_sub(offset),
-            Offsetting::Up => address.wrapping_add(offset),
-        };
-
-        let address: usize = match indexing {
-            Indexing::Pre => effective.try_into().unwrap(),
-            Indexing::Post => address.try_into().unwrap(),
-        };
-
-        let mut mem = self.memory.lock().unwrap();
-
-        match load_store {
-            LoadStoreKind::Store => {
-                let value = if rd_source_destination_register == REG_PROGRAM_COUNTER {
-                    let pc: u32 = self.registers.program_counter().try_into().unwrap();
-                    pc + 12
-                } else {
-                    self.registers
-                        .register_at(rd_source_destination_register as usize)
-                };
-
-                match transfer_type {
-                    HalfwordTransferType::UnsignedHalfwords => {
-                        mem.write_at(address, value.get_bits(0..=7) as u8);
-                        mem.write_at(address + 1, value.get_bits(8..=15) as u8);
-                    }
-                    _ => unreachable!("HS flags can't be != from 01 for STORE (L=0)"),
-                };
-            }
-            LoadStoreKind::Load => match transfer_type {
-                HalfwordTransferType::UnsignedHalfwords => {
-                    let v = mem.read_half_word(address);
-                    self.registers
-                        .set_register_at(rd_source_destination_register as usize, v.into());
-                }
-                HalfwordTransferType::SignedByte => {
-                    let v = mem.read_at(address) as u32;
-                    self.registers.set_register_at(
-                        rd_source_destination_register as usize,
-                        v.sign_extended(8),
-                    );
-                }
-                HalfwordTransferType::SignedHalfwords => {
-                    let v = mem.read_half_word(address) as u32;
-                    self.registers.set_register_at(
-                        rd_source_destination_register as usize,
-                        v.sign_extended(16),
-                    );
-                }
-            },
-        }
-
-        if indexing == Indexing::Post || write_back {
-            self.registers
-                .set_register_at(rn_base_register.try_into().unwrap(), effective);
-        }
-
-        if !(load_store == LoadStoreKind::Load
-            && rd_source_destination_register == REG_PROGRAM_COUNTER)
-        {
-            Some(SIZE_OF_ARM_INSTRUCTION)
-        } else {
-            None
-        }
-    }
-
     pub fn swap_mode(&mut self, new_mode: Mode) {
         if self.cpsr.mode() == new_mode {
             return;
@@ -811,153 +685,6 @@ impl Arm7tdmi {
         }
 
         self.cpsr.set_mode(new_mode);
-    }
-
-    fn branch(&mut self, is_link: bool, offset: u32) -> Option<u32> {
-        let offset = offset.sign_extended(26) as i32;
-        let old_pc: u32 = self.registers.program_counter().try_into().unwrap();
-        if is_link {
-            self.registers
-                .set_register_at(14, old_pc.wrapping_add(SIZE_OF_ARM_INSTRUCTION));
-        }
-
-        // 8 is for the prefetch
-        let new_pc = self.registers.program_counter() as i32 + offset + 8;
-        self.registers.set_program_counter(new_pc as u32);
-
-        // Never advance PC after B
-        None
-    }
-
-    #[allow(clippy::too_many_arguments)]
-    fn block_data_transfer(
-        &mut self,
-        indexing: Indexing,
-        offsetting: Offsetting,
-        load_psr: bool,
-        write_back: bool,
-        load_store: LoadStoreKind,
-        rn: u32,
-        reg_list: u32,
-    ) -> Option<u32> {
-        let base_register = rn.try_into().unwrap();
-        let memory_base = self.registers.register_at(base_register);
-        let mut address = memory_base.try_into().unwrap();
-
-        if load_psr {
-            unimplemented!();
-        }
-
-        let transfer = match load_store {
-            LoadStoreKind::Store => {
-                |arm: &mut Self, address: usize, reg_source: usize| {
-                    let mut value = arm.registers.register_at(reg_source);
-
-                    // If R15 we get the value of the current instruction + 12
-                    if reg_source == REG_PROGRAM_COUNTER.try_into().unwrap() {
-                        value += 12;
-                    }
-                    let mut memory = arm.memory.lock().unwrap();
-
-                    memory.write_at(address, value.get_bits(0..=7) as u8);
-                    memory.write_at(address + 1, value.get_bits(8..=15) as u8);
-                    memory.write_at(address + 2, value.get_bits(16..=23) as u8);
-                    memory.write_at(address + 3, value.get_bits(24..=31) as u8);
-                }
-            }
-            LoadStoreKind::Load => |arm: &mut Self, address: usize, reg_destination: usize| {
-                let memory = arm.memory.lock().unwrap();
-                let part_0: u32 = memory.read_at(address).try_into().unwrap();
-                let part_1: u32 = memory.read_at(address + 1).try_into().unwrap();
-                let part_2: u32 = memory.read_at(address + 2).try_into().unwrap();
-                let part_3: u32 = memory.read_at(address + 3).try_into().unwrap();
-                drop(memory);
-                let v = part_3 << 24_u32 | part_2 << 16_u32 | part_1 << 8_u32 | part_0;
-                arm.registers.set_register_at(reg_destination, v);
-            },
-        };
-
-        self.exec_data_transfer(reg_list, indexing, &mut address, offsetting, transfer);
-
-        if write_back {
-            self.registers
-                .set_register_at(base_register, address.try_into().unwrap());
-        };
-
-        // If LDM and R15 is in register list we don't advance PC
-        if !(load_store == LoadStoreKind::Load && reg_list.is_bit_on(15)) {
-            Some(SIZE_OF_ARM_INSTRUCTION)
-        } else {
-            None
-        }
-    }
-
-    #[allow(clippy::too_many_arguments)]
-    fn coprocessor_data_transfer(
-        &mut self,
-        indexing: Indexing,
-        offsetting: Offsetting,
-        _transfer_length: bool,
-        _write_back: bool,
-        _load_store: LoadStoreKind,
-        rn: u32,
-        _crd: u32,
-        _cp_number: u32,
-        offset: u32,
-    ) -> Option<u32> {
-        let mut _address = self.registers.register_at(rn.try_into().unwrap());
-        let effective = match offsetting {
-            Offsetting::Down => _address.wrapping_sub(offset),
-            Offsetting::Up => _address.wrapping_add(offset),
-        };
-
-        let _address = match indexing {
-            Indexing::Pre => effective,
-            Indexing::Post => _address,
-        };
-
-        // TODO: take a look if we need to finish this for real.
-        todo!("finish this");
-        // Some(SIZE_OF_ARM_INSTRUCTION)
-    }
-
-    fn exec_data_transfer<F>(
-        &mut self,
-        reg_list: u32,
-        indexing: Indexing,
-        address: &mut usize,
-        offsetting: Offsetting,
-        transfer: F,
-    ) where
-        F: Fn(&mut Self, usize, usize),
-    {
-        let alignment = 4; // Since are word, the alignment is 4.
-
-        let change_address = |address: usize| match offsetting {
-            Offsetting::Down => address.wrapping_sub(alignment),
-            Offsetting::Up => address.wrapping_add(alignment),
-        };
-
-        // If we are decreasing we still want to store the lowest reg to the lowest
-        // memory address. For this reason we reverse the loop order.
-        let range_registers: Box<dyn Iterator<Item = u8>> = match offsetting {
-            Offsetting::Down => Box::new((0..=15).rev()),
-            Offsetting::Up => Box::new(0..=15),
-        };
-
-        for reg_source in range_registers {
-            if reg_list.is_bit_on(reg_source) {
-                if indexing == Indexing::Pre {
-                    *address = change_address(*address);
-                }
-
-                transfer(self, *address, reg_source.into());
-
-                if indexing == Indexing::Post {
-                    *address = change_address(*address);
-                }
-            }
-        }
     }
 
     pub fn pc_relative_load(&mut self, r_destination: u16, immediate_value: u16) -> Option<u32> {
