+++ emu/src/cpu/mod.rs
@@ -2,11 +2,9 @@ mod arm;
 pub mod arm7tdmi;
 mod condition;
 mod cpu_modes;
-mod data_processing;
 mod flags;
 mod move_compare_add_sub;
 mod psr;
 mod register_bank;
 mod registers;
-mod single_data_transfer;
 mod thumb;
